+++ CHANGELOG.md
@@ -10,9 +10,9 @@
 - In the overview map when hovering over a borehole, clicking the tooltip text now navigates to the borehole detail view. If several boreholes are located at the same position, a list of boreholes is displayed in the tooltip and clicking on one of them navigates to the respective borehole detail view.
 - Administrators can now edit workgroup names on the settings page.
 - Show loading indicator when importing data.
-- Remove type from profile table.
 - Select input fields now filter and display only the options that match the user's input.
-
+- Remove type from profile table.
+- 
 ### Fixed
 - Boreholes could not be exported in the _view_ environment.
 - `Boreholes.swissgeol.ch ID` was not listed in the location type ID filter.